OpenMP decompression v1.0
This patch adds support for OpenMP decompression. For fixed rate, the
block positions are computed. For variable rate, an index which contains
the offsets w.r.t. the beginning of the compressed stream is required.
This patch support an index with the 'offsets' format, with variable
granularity. The index can be computed from a block length table, which
is generated during compression.